+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - LDAP authenticator: hide secondary email field in account settings when logged in via LDAP
+- Calendar: morkaround for event store when no calendars in search paramaeters
 
 29-06-2026: 6.8.174
 - Core: fix import in APCu cache class